% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_add_info.R
\name{add.player.info}
\alias{add.player.info}
\title{Add additional player information}
\usage{
add.player.info(dt.input)
}
\arguments{
\item{dt.input}{data.frame}
}
\value{
data.frame
}
\description{
Add additional player information
}
