% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{baseball.preprocess}
\alias{baseball.preprocess}
\title{This function pre-processes the names of the baseball players}
\usage{
baseball.preprocess(dt.input)
}
\arguments{
\item{dt.input}{data.frame}
}
\value{
data.frame
}
\description{
This function pre-processes the names of the baseball players
}
