% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{basketball.preprocess}
\alias{basketball.preprocess}
\title{This function pre-processes the names of the basketball players}
\usage{
basketball.preprocess(dt.input)
}
\arguments{
\item{dt.input}{data.frame}
}
\value{
data.frame
}
\description{
This function pre-processes the names of the basketball players
}
