% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{data.preprocess}
\alias{data.preprocess}
\title{This function pre-processes the card title column}
\usage{
data.preprocess(v.player.names)
}
\arguments{
\item{v.player.names}{character}
}
\value{
data.frame
}
\description{
This function pre-processes the card title column
}
