% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_add_info.R
\name{drop.useless.columns}
\alias{drop.useless.columns}
\title{Drop columns that are no longer required}
\usage{
drop.useless.columns(dt.input)
}
\arguments{
\item{dt.input}{data.frame}
}
\value{
data.frame
}
\description{
Drop columns that are no longer required
}
