% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_products_id.R
\name{fix.names}
\alias{fix.names}
\title{There's an issue with some names that contain "jr." for the baseball cards
this function addresses that issue}
\usage{
fix.names(dt.input, sport)
}
\arguments{
\item{dt.input}{data.frame}

\item{sport}{character}
}
\description{
There's an issue with some names that contain "jr." for the baseball cards
this function addresses that issue
}
