% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_products_id.R
\name{flag.id.valid}
\alias{flag.id.valid}
\title{Tags which observations have a valid product id. These are the observations
for which exactly one player has been identified.}
\usage{
flag.id.valid(dt.input)
}
\arguments{
\item{dt.input}{data.frame}
}
\description{
Tags which observations have a valid product id. These are the observations
for which exactly one player has been identified.
}
