% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player.R
\name{get.player}
\alias{get.player}
\title{Extract player name(s)}
\usage{
get.player(dt.input, card.title.col.name = "card_title", sport)
}
\arguments{
\item{dt.input}{data.frame}

\item{card.title.col.name}{character}

\item{sport}{character}
}
\value{
data.frame
}
\description{
Use a card's title to extract the player/players name/names that it refers to
Returns a data frame consisting of the (1) input card title, (2) number of
identified players, (3) name(s) of the identified player(s) (each name in a
different column)
}
\examples{
dt.example.basketball.names <- data.table::rbindlist(
apply(dt.example.basketball.names, 1, get.player,
card.title.col.name = "card_title", sport = "basketball"), fill=TRUE)
dt.example.baseball.names <- data.table::rbindlist(
apply(dt.example.baseball.names, 1, get.player,
card.title.col.name = "card_title", sport = "baseball"), fill=TRUE)
}
