% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year.R
\name{get.product.year}
\alias{get.product.year}
\title{Use the extracted years from get.year.from.title and
get.year.from.product.section to construct a card year column}
\usage{
get.product.year(dt.input)
}
\arguments{
\item{dt.input}{data.frame}
}
\value{
data.frame
}
\description{
Use the extracted years from get.year.from.title and
get.year.from.product.section to construct a card year column
}
