% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year.R
\name{get.year.from.product.section}
\alias{get.year.from.product.section}
\title{Extract card year from product section}
\usage{
get.year.from.product.section(
  dt.input,
  product.section.year.col.name = "product_section_year"
)
}
\arguments{
\item{dt.input}{data.frame}

\item{product.section.year.col.name}{character}
}
\value{
data.frame
}
\description{
Use a card's product section information to extract the year that it refers to
Returns a data frame consisting of the (1) input card title, (2) a set of
columns with identified years. Product section information is commonly used
in marketplaces such as eBay. Sellers of items may or may not fill in this
information.
}
\examples{
dt.example.year <- rbindlist(apply(dt.example.year, 1, get.year.from.product.section,
product.section.year.col.name = "product_section_year"), fill=TRUE)
}
