% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year.R
\name{get.year.from.title}
\alias{get.year.from.title}
\title{Extract card year from card title}
\usage{
get.year.from.title(dt.input, card.title.col.name = "card_title")
}
\arguments{
\item{dt.input}{data.frame}

\item{card.title.col.name}{character}
}
\value{
data.frame
}
\description{
Use a card's title to extract the year that it refers to
Returns a data frame consisting of the (1) input card title, (2) a set of
columns with identified years
}
