% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_products_id.R
\name{player.tie.break}
\alias{player.tie.break}
\title{In some cases the identified player names can refer to players who were
active during different periods in time (e.g., Bob Smith for baseball). This
function tries to address this issue by utilizing the identified card year
and pinpointing which specific player each card refers to.}
\usage{
player.tie.break(dt.input, sport)
}
\arguments{
\item{dt.input}{data.frame}

\item{sport}{character}
}
\description{
In some cases the identified player names can refer to players who were
active during different periods in time (e.g., Bob Smith for baseball). This
function tries to address this issue by utilizing the identified card year
and pinpointing which specific player each card refers to.
}
