% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_products_id.R
\name{remove.duplicate.players}
\alias{remove.duplicate.players}
\title{For some entries the previous functions have identified identical player
names that have been inserted in separate columns. The below function
removes the duplicates and only keeps the unique names}
\usage{
remove.duplicate.players(dt.input, sport)
}
\arguments{
\item{dt.input}{data.frame}

\item{sport}{character}

\item{card.title.col.name}{character}
}
\description{
For some entries the previous functions have identified identical player
names that have been inserted in separate columns. The below function
removes the duplicates and only keeps the unique names
}
